^ [#22786] Merge revision 4554 from branches/2.0-xillibit-mvc-backend-26-02-2011
git-svn-id: file:///Users/oratzes/svn/kunena@4584 daf2aa89-cb5b-0410-b8ee-fe93ebc9b27d

+++ trunk/2.0/administrator/components/com_kunena/controllers/prune.php
@@ -57,12 +57,14 @@ class KunenaAdminControllerPrune extends KunenaController {
 		$deleted = 0;
 		foreach ( $topics as $topic ) {
 			$deleted++;
-			$topic->delete(false);
+			if ( $trashdelete ) $topic->delete(false);
+			else $topic->trash();
 		}
 		KunenaUserHelper::recount();
 		KunenaForumCategoryHelper::recount();
 		KunenaForumMessageAttachmentHelper::cleanup();
-		$app->enqueueMessage ( "" . JText::_('COM_KUNENA_FORUMPRUNEDFOR') . " " . $prune_days . " " . JText::_('COM_KUNENA_PRUNEDAYS') . "; " . JText::_('COM_KUNENA_PRUNEDELETED') . " {$deleted}/{$count} " . JText::_('COM_KUNENA_PRUNETHREADS') );
+		if ( $trashdelete ) $app->enqueueMessage ( "" . JText::_('COM_KUNENA_FORUMPRUNEDFOR') . " " . $prune_days . " " . JText::_('COM_KUNENA_PRUNEDAYS') . "; " . JText::_('COM_KUNENA_PRUNEDELETED') . " {$deleted}/{$count} " . JText::_('COM_KUNENA_PRUNETHREADS') );
+		else $app->enqueueMessage ( "" . JText::_('COM_KUNENA_FORUMPRUNEDFOR') . " " . $prune_days . " " . JText::_('COM_KUNENA_PRUNEDAYS') . "; " . JText::_('COM_KUNENA_PRUNETRASHED') . " {$deleted}/{$count} " . JText::_('COM_KUNENA_PRUNETHREADS') );
 		$this->setRedirect(KunenaRoute::_($this->baseurl, false));
 	}
 }
